Template: vite-cfagents-runner
Base template for I want to build a page dedicated to beagles, I Want to be able to upload images, bot individually and in bulk, directly from the photos app on my iphone. I Aslo want to be able to buy dogfood, dog toys, and grooming stuff. I walso want to be able to book a grooming session as well a get vetenarian advice specific to Beagles